+++ lib/chemical_equation_balancer.rb
@@ -1,0 +1,4 @@
+
+Dir.glob("#{__dir__}/chemical_equation_balancer/*.rb").each do |filename|
+  require filename
+end
